Move `util.js` back to `lib/`.
Util doesn’t need to be access directly:

- it’s available as `targaryen.util`
- No circular dependencies; plugins can import access them via `targaryen.util`;
