+++ lib/util.js
@@ -1,8 +1,8 @@
 
 'use strict';
 
-const ruleset = require('./lib/ruleset');
-const database = require('./lib/database');
+const ruleset = require('./ruleset');
+const database = require('./database');
 
 let debug = true;
 let data, rules, db;
